sync-github-releases: split the API client from environment resolution
utils/github.ts mixed two concerns: the GitHub Releases API client (pure
transport) and the resolution of which repository / token / default branch
to act on (read from the environment). The latter even pulled a `git`
import into the transport file.

Move getGithubToken/getDefaultBranch/getRepository into utils/github-env.ts,
leaving github.ts as the client alone — now free of any git dependency.
